Fix: SQLite Sum on decimal not supported, cast to double

+++ Controllers/DashboardController.cs
@@ -17,7 +17,7 @@ public class DashboardController : Controller
         ViewBag.TotalProducts = await _db.Products.CountAsync();
         ViewBag.TotalClients = await _db.Clients.CountAsync();
         ViewBag.TotalInvoices = await _db.Invoices.CountAsync();
-        ViewBag.TotalRevenue = await _db.Invoices.SumAsync(i => i.Total);
+        ViewBag.TotalRevenue = (await _db.Invoices.SumAsync(i => (double)i.Total));
         ViewBag.ActivePage = "dashboard";
         return View();
     }
